рефакторинг, чтобы запустить приложение
