+++ src/components/add-review/add-review.connect.jsx
@@ -1,7 +1,8 @@
 import AddReview from "@components/add-review/add-review";
 import {connect} from "react-redux";
 import {selectFilmById} from "../../utils/selectors";
-import {Operation} from "../../reducer/review/review";
+// import {Operation} from "../../reducer/review/review";
+import {Operation} from "@redux/reducers/review/review";
 
 
 const mapStateToProps = (state, props) => ({
