+++ src/components/film-details/film-details.connect.jsx
@@ -1,8 +1,10 @@
 import FilmDetails from "@components/film-details/film-details";
 import {connect} from "react-redux";
 import {selectFilmById} from "../../utils/selectors";
-import {Operation} from "../../reducer/reviews/reviews";
-import {Operation as OperationFavorite} from "redux/favorite/favorite";
+// import {Operation} from "../../reducer/reviews/reviews";
+// import {Operation as OperationFavorite} from "redux/favorite/favorite";
+import {Operation as OperationFavorite} from "@redux/reducers/favorite/favorite";
+import {Operation as OperationReviews} from "@redux/reducers/reviews/reviews";
 
 const mapStateToProps = (state, props) => ({
   details: selectFilmById(Number(props.match.params.id), state),
@@ -11,7 +13,7 @@ const mapStateToProps = (state, props) => ({
 
 const mapDispatchToProps = (dispatch) => ({
   getReviews: (id) => {
-    dispatch(Operation.loadReviews(id));
+    dispatch(OperationReviews.loadReviews(id));
   },
   changeStatusFilm: (id, status) => dispatch(OperationFavorite.postFavoriteFilm(id, status))
 });
