+++ src/components/private-route/private-route.jsx
@@ -1,5 +1,6 @@
 import {Route, Redirect} from "react-router";
-import {AuthorizationStatus} from "redux/user/constants/constants.ts";
+// import {AuthorizationStatus} from "redux/user/constants/constants.ts";
+import {AuthorizationStatus} from "@redux/reducers/user/constants/constants";
 import {RoutePathes} from "../../constants/constants.ts";
 
 const PrivateRoute = (props) => {
