+++ src/components/promo-film/promo-film.connect.jsx
@@ -1,6 +1,7 @@
 import PromoFilm from "@components/promo-film/promo-film";
 import {connect} from "react-redux";
-import {Operation as OperationFavorite} from "redux/favorite/favorite";
+// import {Operation as OperationFavorite} from "redux/favorite/favorite";
+import {Operation as OperationFavorite} from "@redux/reducers/favorite/favorite";
 
 const mapDispatchToProps = (dispatch) => ({
   changeStatusFilm: (id, status) => dispatch(OperationFavorite.postFavoriteFilm(id, status))
