+++ src/components/sign-in/sign-in.connect.jsx
@@ -1,6 +1,7 @@
 import {connect} from "react-redux";
 import {SignIn} from "./sign-in";
-import {Operation} from "../../reducer/user/user";
+// import {Operation} from "../../reducer/user/user";
+import {Operation} from "@redux/reducers/user/user";
 
 const mapStateToProps = (state) => ({
   authorizationStatus: state.USER.authorizationStatus
