--- src/reducer/constants/constants.js
@@ -1,11 +1,0 @@
-const NameSpace = {
-  DATA: `DATA`,
-  PROMO: `PROMO`,
-  GENRE: `GENRE`,
-  USER: `USER`,
-  REVIEW: `REVIEW`,
-  REVIEWS: `REVIEWS`,
-  FAVORITE: `FAVORITE`
-};
-
-export default NameSpace;
