--- src/reducer/data/constants/constants.js
@@ -1,8 +1,0 @@
-const ActionType = {
-  LOAD_FILMS_REQUEST: `LOAD_FILMS_REQUEST`,
-  LOAD_FILMS_SUCCESS: `LOAD_FILMS_SUCCESS`,
-  LOAD_FILMS_ERROR: `LOAD_FILMS_ERROR`,
-  CHANGE_FILMS: `CHANGE_FILMS`
-};
-
-export default ActionType;
