--- src/reducer/favorite/constants/constants.js
@@ -1,7 +1,0 @@
-const ActionType = {
-  LOAD_FAVORITE_FILMS_REQUEST: `LOAD_FAVORITE_FILMS_REQUEST`,
-  LOAD_FAVORITE_FILMS_SUCCESS: `LOAD_FAVORITE_FILMS_SUCCESS`,
-  LOAD_FAVORITE_FILMS_ERROR: `LOAD_FAVORITE_FILMS_ERROR`
-};
-
-export default ActionType;
