--- src/reducer/genre/constants/constants.js
@@ -1,5 +1,0 @@
-const ActionType = {
-  CHANGE_ACTIVE_GENRE: `CHANGE_ACTIVE_GENRE`
-};
-
-export default ActionType;
