--- src/reducer/promo/constants/constants.js
@@ -1,8 +1,0 @@
-const ActionType = {
-  LOAD_PROMO_FILM_REQUEST: `LOAD_PROMO_FILM_LOADING`,
-  LOAD_PROMO_FILM_SUCCESS: `LOAD_PROMO_FILM_SUCCESS`,
-  LOAD_PROMO_FILM_ERROR: `LOAD_PROMO_FILM_ERROR`,
-  CHANGE_PROMO_FILM_FIELD: `CHANGE_PROMO_FILM_FIELD`
-};
-
-export default ActionType;
