--- src/reducer/reviews/constants/constants.js
@@ -1,7 +1,0 @@
-const ActionType = {
-  LOAD_REVIEWS_REQUEST: `LOAD_REVIEWS_REQUEST`,
-  LOAD_REVIEWS_SUCCESS: `LOAD_REVIEWS_SUCCESS`,
-  LOAD_REVIEWS_ERROR: `LOAD_REVIEWS_ERROR`
-};
-
-export default ActionType;
