--- src/reducer/user/actions/action.test.js
@@ -1,16 +1,0 @@
-import ActionCreator from "./actions";
-import {ActionType} from "../constants/constants.js";
-import {AuthorizationStatus} from "../constants/constants.js";
-
-describe(`User action creators work correctly`, () => {
-  it(`Action creator for require authorization should return correct action`, () => {
-    expect(ActionCreator.authorizationSuccess(AuthorizationStatus.NO_AUTH)).toEqual({
-      type: ActionType.AUTHORIZATION_SUCCESS,
-      payload: AuthorizationStatus.NO_AUTH
-    });
-    expect(ActionCreator.authorizationSuccess(AuthorizationStatus.AUTH)).toEqual({
-      type: ActionType.AUTHORIZATION_SUCCESS,
-      payload: AuthorizationStatus.AUTH
-    });
-  });
-});
