--- src/reducer/user/utils/adapter.js
@@ -1,10 +1,0 @@
-export default class Adapter {
-  static convertData(data) {
-    return {
-      id: data.id,
-      email: data.email,
-      name: data.name,
-      avatarUrl: data.avatar_url
-    };
-  }
-}
